Move sinon schema matchers and tests into seperate files
